+++ src/main/java/de/theidler/create_mobile_packages/entities/robo_entity/RoboBeeState.java
@@ -6,6 +6,7 @@ public enum RoboBeeState {
     NAVIGATE_TO_TARGET,
     ALIGN_FOR_DELIVERY,
     DELIVER_PACKAGE,
+    PICKUP_PACKAGE,
     LAND,
     SHUTDOWN
 }