+++ src/main/java/de/theidler/create_mobile_packages/index/CMPGuiTextures.java
@@ -17,7 +17,9 @@ public enum CMPGuiTextures implements ScreenElement, TextureSheetSegment {
     PLAYER_NETWORKS_EDIT_NAME("player_networks", 230, 3, 13, 13),
     PLAYER_NETWORKS_BG("player_networks", 2, 36, 210, 22),
     PLAYER_NETWORKS_FOOTER("player_networks", 2, 73, 218, 31),
-    PLAYER_NETWORKS_SLOT("player_networks", 2, 113, 210, 18)
+    PLAYER_NETWORKS_SLOT("player_networks", 2, 113, 210, 18),
+
+    TRASH_MENU("trash_menu", 0, 0, 256, 95)
 
     ;