+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 ------------------------------------------------------
+Create: Mobile Packages - v0.7.0 - 1.21.1 - unreleased
+------------------------------------------------------
+
+### Additions
+
+- Added TrashSlots to the Portable Stock Ticker allowing players to send items (#299)
+    - Set an address
+    - Put items in the TrashSlots
+    - A Robo Bee will come and pick up the items and send them to the address
+
+------------------------------------------------------
 Create: Mobile Packages - v0.6.1 - 1.21.1 - 20.02.2026
 ------------------------------------------------------